requirements.txt
Sphinx
pep8
pyflakes
Django>=1.6,<1.7
django-nose
gdata==2.0.18
xlrd==0.7.1
django-dirtyfields==0.1
django-historicalrecords==1.1
iso8601
south==1.0
gunicorn==0.12.1
supervisor
django-geckoboard
python-memcached==1.48
raven
pytz
django-markup-deprecated
django-autocomplete-light
requests
psycopg2
celery==3.1.12
django-celery==3.1.10
django-celery-email==1.0.4
redis
django-hstore==1.3.5
django-queryset-csv
djangorestframework
django-filter